+++ src/main/kotlin/com/xunfos/ktorPlayground/thrift/basic/AsyncThriftHandler.kt
@@ -1,4 +1,4 @@
-package com.xunfos.ktorPlayground.thrift
+package com.xunfos.ktorPlayground.thrift.basic
 
 import com.xunfos.ktorPlayground.util.traceLog
 import com.xunfos.playground.thrift.GetUserRequest
@@ -6,24 +6,28 @@ import com.xunfos.playground.thrift.GetUserResponse
 import com.xunfos.playground.thrift.GetUsersResponse
 import com.xunfos.playground.thrift.PlaygroundService
 import com.xunfos.playground.thrift.User
+import org.apache.thrift.async.AsyncMethodCallback
 import kotlin.random.Random
 
-class ThriftHandler : PlaygroundService.Iface {
+class AsyncThriftHandler : PlaygroundService.AsyncIface {
     private val rng = Random(System.currentTimeMillis())
 
-    override fun doWork() {
+    override fun doWork(resultHandler: AsyncMethodCallback<Void>) {
         val workId = rng.nextLong()
 
         traceLog("[workId: $workId] Starting doWork Fun. ${WORK_TIME}ms of work")
         Thread.sleep(WORK_TIME)
         traceLog("[workId: $workId] Finishing doWork Fun.")
+
+        resultHandler.onComplete(null)
     }
 
-    override fun ping() {
+    override fun ping(resultHandler: AsyncMethodCallback<Void>) {
         traceLog("Pong!")
+        resultHandler.onComplete(null)
     }
 
-    override fun getUser(request: GetUserRequest): GetUserResponse {
+    override fun getUser(request: GetUserRequest, resultHandler: AsyncMethodCallback<GetUserResponse>) {
         val workId = rng.nextLong()
 
         traceLog("[workId: $workId] Starting getUser Fun. ${WORK_TIME}ms of work")
@@ -31,28 +35,30 @@ class ThriftHandler : PlaygroundService.Iface {
         traceLog("[workId: $workId] Finishing getUser Fun.")
 
         // Suppose a query would be made
-        return GetUserResponse().apply {
-            this.user = User().apply {
-                this.id = request.id
-                this.name = "banana"
+        resultHandler.onComplete(
+            GetUserResponse().apply {
+                this.user = User().apply {
+                    this.id = request.id
+                    this.name = "banana"
+                }
             }
-        }
+        )
     }
 
-    override fun getUsers(): GetUsersResponse {
+    override fun getUsers(resultHandler: AsyncMethodCallback<GetUsersResponse>) {
         val workId = rng.nextLong()
 
         traceLog("[workId: $workId] Starting getUsers Fun. ${WORK_TIME}ms of work")
         Thread.sleep(WORK_TIME)
         traceLog("[workId: $workId] Finishing getUsers Fun.")
 
-        return GetUsersResponse().apply {
+        resultHandler.onComplete(GetUsersResponse().apply {
             this.users = listOf(
                 User().apply {
                     this.id = "1"
                     this.name = "banana"
                 })
-        }
+        })
     }
 
     companion object {
